improva data bus

+++ data-bus/src/data_bus_player.rs
@@ -12,7 +12,7 @@ impl DataBusPlayer {
     /// # Arguments
     /// * `data_bus` - The `DataBus` to which the data is sent.
     /// * `data` - A vector of `(BusId, Payload)` tuples.
-    pub fn play<D, BD: BusDevice<D>>(data_bus: &mut DataBus<D, BD>, data: Vec<(u16, Vec<D>)>) {
+    pub fn play<D, BD: BusDevice<D>>(data_bus: &mut DataBus<D, BD>, data: Vec<(usize, Vec<D>)>) {
         for (bus_id, payload) in data {
             data_bus.write_to_bus(bus_id, &payload);
         }
